Use Artifacts directly
This works around our lack of cross-targeting in JLLWrappers by instead
building an `Artifacts.toml` file manually that copies the stanzas from
the actual JLL release, but duplicates the `x86_64` stanzas to `i686`
stanzas.  This means that if we were to actually run this on true `i686`
hardware it would fail, but that is hopefully much less common than
running `i686` binaries on `x86_64` hardware.

We maintain rough API compatibility with JLLWrappers in expectation of
being able to generate a JLL with the appropriate `Artifacts.toml` in
the near future.